Fix reading ITK center of rotation
Since the actual application of the center of rotation is well-tested,
the only possible cause of errors would be the LPS → RAS conversion.  So
do that separately at the end.

Prior to this commit, there was definitely a broken interaction between
the center of rotation and the LPS → RAS conversion.